Resolved bug 893293
Current state of code:
Whirlpools and storms swallow appropriate monsters but will not touch each other.
It's hard to recreate in the old systems, but it seems reasonable for monsters to be affected by forces of nature both whirlpool and twister alike.
Comment on this bug to discuss further.

(related issues were affected when adding the phantoms of the 4th floor of destard, which the party could walk on)

git-svn-id: svn://svn.code.sf.net/p/xu4/code/trunk@2764 694b2c84-8326-496a-a978-df44a0d360f4