Fix warnings.
I know it's a minor thing, but it's kind of nice to have things compile without
warnings. This should make it easier to find other issues that show up. That
is, if you see a warning now, it's likely that it is new code that added it.

This corresponds to patch 3196483 (with two additions).

git-svn-id: svn://svn.code.sf.net/p/xu4/code/trunk@2844 694b2c84-8326-496a-a978-df44a0d360f4