Fix up some bugs with the Intro Screen on iPad.
The timing and other things weren't being showed correctly on the iPad. The
timing was incredibly off. Now it should do the right thing. It also is smart
enough to restart the timer from the accelerated version that is used for
drawing the signature.

There are still things messed up with the intro screen, (e.g., the title
doesn't fade in and the map is shown incorectly when the "scroll" is rolled
out) but it is much closer.

git-svn-id: svn://svn.code.sf.net/p/xu4/code/trunk@2892 694b2c84-8326-496a-a978-df44a0d360f4