A better fix for the crash fix I submitted last night.
This is a compromise and probably the best without a disruptive architecture
change. It's only the combatcontroller that is causing the problem by deleting
itself, so my sleep now runs the event loop like it was before, but during
combat, we pause input. That way, the thread doesn't sleep and make potentially
lock up the system.

git-svn-id: svn://svn.code.sf.net/p/xu4/code/trunk@2898 694b2c84-8326-496a-a978-df44a0d360f4