Unify screen functions between iOS and SDL.
I had copied a lot of functionality from the SDL source code for iOS, but most
of it was just copy and paste and most of it was "cross platform," that the
source code did not address any specific iOS and SDL functions. The upshot was
that we would lose any updates from one file that wasn't in another. I've now
put all the shared code in the screen.cpp. This should make it much easier to
fix the generic screen functions on multiple platforms.

I'm not sure this current function-based version is the best design, it
probably should be a class, but that is too large a change for the moment.
idea here is to reduce bug-fixes. For example, peering at gems on iOS functions
exactly the same.

git-svn-id: svn://svn.code.sf.net/p/xu4/code/trunk@2908 694b2c84-8326-496a-a978-df44a0d360f4