Fix Lord British's farewell.
In the original game, Lord British would say "friend" or "friends" depending on
how many people in the party you have. We now do this. It's a bit heavy handed
in that we unload the entire level, but on the other side, it paves the way for
doing unloading of cities to save memory (if that is needed).

I also found a bug in the reference counting where we would decrement twice.

git-svn-id: svn://svn.code.sf.net/p/xu4/code/trunk@2994 694b2c84-8326-496a-a978-df44a0d360f4