Fix up crashing the other superbutton helper.
Argh. Sometimes it seems that my C++ inheritance stuff is rusty. Of course,
subclasses get their parent class constructor called. Honestly, the small
amount of code I was trying to save (already put in a function) is not worth
the headache.Anyway, this should stop crashes when in the dungeon.

git-svn-id: svn://svn.code.sf.net/p/xu4/code/trunk@3007 694b2c84-8326-496a-a978-df44a0d360f4