Fix compilation to work on iOS again.
I understand the optimizations that were done a little while ago seemed like a
good idea, but they broke the iOS version. Sometimes the abstraction is a good
thing. Also, the places that I reverted them back (initializing and turning on
and off music) are not bottlenecks in the code.

git-svn-id: svn://svn.code.sf.net/p/xu4/code/trunk@3032 694b2c84-8326-496a-a978-df44a0d360f4